Chore:: Refactor web3 donations  and removed all chains that are not yet supported by the donation smart contracts
