+++ src/constants/contract-address/types.ts
@@ -1,0 +1,1 @@
+export type DonationToken = "USDT" | "USDC";
